move more functions from update into separate JS files
